backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.63.0"
uvicorn = {extras = ["standard"], version = "^0.13.4"}
SQLAlchemy = "^1.4.6"
alembic = "^1.5.8"
celery = "^5.0.5"
python-decouple = "^3.4"
psycopg2 = "^2.8.6"
passlib = "^1.7.4"
python-jose = {extras = ["cryptography"], version = "^3.2.0"}
python-multipart = "^0.0.5"
aioredis = "^1.3.1"
requests = "^2.25.1"
pyppeteer = "^0.2.5"
lxml = "^4.6.3"

[tool.poetry.dev-dependencies]
mypy = "^0.812"
black = "^20.8b1"
flake8 = "^3.9.0"
pytest = "^6.2.4"
pytest-asyncio = "^0.15.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
